# Conflicts:
#	platforms/common/src/CommonKernels.cpp

+++ platforms/common/include/openmm/common/ArrayInterface.h
@@ -9,7 +9,7 @@
  * Biological Structures at Stanford, funded under the NIH Roadmap for        *
  * Medical Research, grant U54 GM072970. See https://simtk.org.               *
  *                                                                            *
- * Portions copyright (c) 2019-2021 Stanford University and the Authors.      *
+ * Portions copyright (c) 2019-2022 Stanford University and the Authors.      *
  * Authors: Peter Eastman                                                     *
  * Contributors:                                                              *
  *                                                                            *
@@ -51,7 +51,7 @@ public:
      * @param elementSize       the size of each element in bytes
      * @param name              the name of the array
      */
-    virtual void initialize(ComputeContext& context, int size, int elementSize, const std::string& name) = 0;
+    virtual void initialize(ComputeContext& context, size_t size, int elementSize, const std::string& name) = 0;
     /**
      * Initialize this object.  The template argument is the data type of each array element.
      *
@@ -60,13 +60,13 @@ public:
      * @param name              the name of the array
      */
     template <class T>
-    void initialize(ComputeContext& context, int size, const std::string& name) {
+    void initialize(ComputeContext& context, size_t size, const std::string& name) {
         initialize(context, size, sizeof(T), name);
     }
     /**
      * Recreate the internal storage to have a different size.
      */
-    virtual void resize(int size) = 0;
+    virtual void resize(size_t size) = 0;
     /**
      * Get whether this array has been initialized.
      */
@@ -74,7 +74,7 @@ public:
     /**
      * Get the number of elements in the array.
      */
-    virtual int getSize() const = 0;
+    virtual size_t getSize() const = 0;
     /**
      * Get the size of each element in bytes.
      */
